# 	improve the performance of fs.Open() by unzipping in fs.New() and simplifying http.File wrapper