Please consider reviewing and merging https://github.com/rakyll/statik/pull/22 before this PR. I'd like to rebase this branch after that PR is merged to take advantage of the additional test coverage.

Benchmark results (from benchmark in https://github.com/rakyll/statik/pull/22):
master:
```
$ go test -v -bench=. -run=^$ github.com/rakyll/statik/fs
BenchmarkOpen-4   	  500000	      2398 ns/op
PASS
ok  	github.com/rakyll/statik/fs	1.237s
```
this PR:
```
$ go test -v -bench=. -run=^$ github.com/rakyll/statik/fs
BenchmarkOpen-4   	10000000	       137 ns/op
PASS
ok  	github.com/rakyll/statik/fs	1.526s
```